--- DoublePendulum/index.html
@@ -1,51 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <script
-      language="javascript"
-      type="text/javascript"
-      src="../libraries/p5.min.js"
-    ></script>
-    <script type="module">
-      import { sketch } from "./DoublePendulum.js";
-      import "../sketchlib/sketchbook-navbar.js";
-      const p5_sketch = new p5(sketch, "sketch");
-    </script>
-    <link rel="stylesheet" href="../styles/layout.css" />
-    <script
-      id="MathJax-script"
-      async
-      src="https://cdn.jsdelivr.net/npm/mathjax@3/es5/tex-mml-chtml.js"
-    ></script>
-  </head>
-  <body>
-    <sketchbook-navbar></sketchbook-navbar>
-    <div class="container">
-      <h1>Double Pendulum (2016, 2024)</h1>
-      <div class="sketch">
-        <canvas id="sketch-canvas" width="500" height="700"></canvas>
-      </div>
-      <div class="description">
-        <!-- prettier-ignore -->
-        <p>
-        
-
-        $$
-        \newcommand{\dth}[]{\dot{\theta}}
-        \newcommand{\ddth}[]{\ddot{\theta}}
-
-
-        \begin{align}
-            M &= m_1 + m_2 \\
-            a &= \theta_1 - \theta_2 \\
-
-            \dth_1 &= \dth_1 \\
-            \ddth_1 &= \frac{L_1 m_2 \ddth_1^2 \sin(a)\cos(a)}{B} \\
-        \end{align}$$
-        </p>
-      </div>
-    </div>
-  </body>
-</html>
