+++ lab/DoublePendulum/index.html
@@ -1,0 +1,74 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <script
+      language="javascript"
+      type="text/javascript"
+      src="../../libraries/p5.min.js"
+    ></script>
+    <script type="module">
+      import { sketch } from "./DoublePendulum.js";
+      import "../../sketchlib/sketchbook-navbar.js";
+      const p5_sketch = new p5(sketch, "sketch");
+    </script>
+    <link rel="stylesheet" href="../../styles/layout.css" />
+    <script
+      id="MathJax-script"
+      async
+      src="https://cdn.jsdelivr.net/npm/mathjax@3/es5/tex-mml-chtml.js"
+    ></script>
+  </head>
+  <body>
+    <sketchbook-navbar></sketchbook-navbar>
+    <div class="container">
+      <h1>Double Pendulum (2016, 2025)</h1>
+      <div class="sketch">
+        <canvas id="sketch-canvas" width="500" height="700"></canvas>
+      </div>
+      <div class="description">
+        Description Outline:
+        <ul>
+          <li>
+            Mention the history - This was originally a Processing.py sketch I
+            made in college.
+          </li>
+          <li>I was learning about chaotic motion and Lagrangian dynamics</li>
+          <li>Link to <a href="../DoubleSpring/">Double Spring</a></li>
+          <li>
+            mention
+            <a href="https://youtu.be/3VqbW6H_ets">The YouTube video</a> I made
+            of the original version
+          </li>
+          <li>
+            I think I had some details written up in an old iteration of my
+            website? I'll have to look through the git history for that...
+          </li>
+          <li>
+            I do want to re-derive the equations... but that's best done in a
+            README file since it's rather technical. Maybe make a high-level
+            description of how the forces lead to chaotic motion?
+          </li>
+        </ul>
+        <!-- prettier-ignore -->
+        <!--
+        <p>
+        $$
+        \newcommand{\dth}[]{\dot{\theta}}
+        \newcommand{\ddth}[]{\ddot{\theta}}
+
+
+        \begin{align}
+            M &= m_1 + m_2 \\
+            a &= \theta_1 - \theta_2 \\
+
+            \dth_1 &= \dth_1 \\
+            \ddth_1 &= \frac{L_1 m_2 \ddth_1^2 \sin(a)\cos(a)}{B} \\
+        \end{align}$$
+        </p>
+        -->
+      </div>
+    </div>
+  </body>
+</html>
