+++ lab/DoubleSpring/index.html
@@ -18,11 +18,20 @@
   <body>
     <sketchbook-navbar></sketchbook-navbar>
     <div class="container">
-      <h1>TITLE (YEAR)</h1>
+      <h1>Double Spring (2016, 2025)</h1>
       <div class="sketch">
         <canvas id="sketch-canvas" width="500" height="700"></canvas>
       </div>
-      <div class="description">DESCRIPTION GOES HERE</div>
+      <div class="description">
+        <ul>
+          <li>Link to <a href="../DoubleSpring/">Double Spring</a></li>
+          <li>
+            Mention
+            <a href="https://youtu.be/GoOVCgcw65M">the YouTube video </a> I made
+            from my original version
+          </li>
+        </ul>
+      </div>
     </div>
   </body>
 </html>
