+++ lab/index.html
@@ -1,0 +1,55 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="stylesheet" href="../styles/layout.css" />
+    <style>
+      .linklist {
+        width: 100%;
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        gap: 10px;
+      }
+
+      .link {
+        width: 50%;
+        max-width: 600px;
+        background-color: var(--background-light);
+        padding: 10px;
+        border-radius: 10px;
+        display: flex;
+        flex-direction: row;
+        align-items: center;
+        flex-wrap: wrap;
+        gap: 10px;
+      }
+
+      @media screen and (max-width: 500px) {
+        .link {
+          width: 250px;
+        }
+      }
+    </style>
+    <script type="module" src="../sketchlib/sketchbook-navbar.js"></script>
+  </head>
+
+  <body>
+    <sketchbook-navbar></sketchbook-navbar>
+    <div id="container" class="linklist">
+      <h1>P5 Sketchbook Lab🧪</h1>
+      <p>
+        I wanted to make a place to put my work-in-progress art experiments.
+        This way I can share things with friends sooner for feedback. They may
+        be rough around the edges compared to the
+        <a href="../">main gallery</a>. This list will change over time, so
+        check back often!
+      </p>
+      <ul>
+        <li><a href="./DoublePendulum/">Double Pendulum</a></li>
+        <li><a href="./DoubleSpring/">Double Spring</a></li>
+      </ul>
+    </div>
+  </body>
+</html>
